feature: adds support for notice script and stylesheet
